requirements.txt
streamlit
pandas
scikit-learn==0.24.2
pillow==8.3.1
